feat: fix auth route
