+++ src/app/favorites/page.js
@@ -1,4 +1,4 @@
-import { authOptions } from '@/app/api/auth/[...nextauth]/route'
+import { authOptions } from '@/app/lib/auth'
 import Card from '@/components/Card'
 import { getServerSession } from 'next-auth'
